flask_apps/app/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyzbar==0.1.9
requests==2.32.3
segno==1.6.1
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
